requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
bulletchess==0.1.9
python-chess>=1.999